+++ src/pages/mypage/addRecipePage/component/addRecipeIngredients.jsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import style from "../../css/recipe/addRecipeIngredients.module.css"
+import style from "../style/addRecipeIngredients.module.css"
 import { BiSolidMinusCircle } from "react-icons/bi";
 
 export default function AddRecipeIngredients({ingredients}) {
