+++ src/pages/mypage/addRecipePage/style/addrecipepage.module.css
@@ -1,0 +1,100 @@
+.frame{
+    width: 500px;
+    padding: 20px 0px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    border-radius: 10px;
+    border: 1px solid orange;
+}
+
+.foodInfo{
+    width: 700px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    font-size: 14px;
+}
+
+.foodInfo_left{
+    width: 375px;
+}
+
+.inputBox{
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    width: 100%;
+}
+
+.inputBox p{
+    color: orange;
+    width: 140px;
+}
+
+.inputBox input{
+    width: 100%;
+    border-radius: 6px;
+    border: 1px solid orange;
+}
+
+.inputBox textarea{
+    width: 100%;
+    height: 100px;
+    resize: none;
+    border-radius: 10px;
+    border: 1px solid orange;
+}
+
+.selectBox{
+    display: flex;
+    align-items: center;
+    justify-content: flex-start;
+}
+
+.selectBox p{
+    color: orange;
+    width: 80px;
+    text-align: right;
+}
+
+.selectBox select{
+    width: 130px;
+    border-radius: 6px;
+    border: 1px solid orange;
+    margin: 0px 15px;
+}
+
+hr {
+    width: 85%;
+    border: none;
+    height: 0px;
+    border-top: 2px solid #ffbb00c2;
+    margin: 30px 0px;
+  }
+
+  .tagBox{
+    display: flex;
+    align-items: flex-start;
+    justify-content: space-between;
+    width: 100%;
+    flex-direction: column;
+}
+
+.buttonBox{
+    display: flex;
+    width: 700px;
+    height: 80px;
+    align-items: center;
+    justify-content: center;
+    margin-top: 70px;
+}
+
+.saveButton{
+    width: 60%;
+    height: 63px;
+    border-radius: 10px;
+    background-color: #ff9e01;
+    color: white;
+    font-size: 20px;
+}
